virt: test: rename TestingVm -> FakeVm
We have a Fake in vm_test.py named TestingVm, and py.test's
autodetection believes this is a testsuite. But since the class
also have __init__defined, it is discarded (see
https://docs.pytest.org/en/latest/goodpractices.html#conventions-for-python-test-discovery)

The only effect of this is a slightly confusing warning emitted
when out testsuite ends.

This patch renames the 'TestingVm' helper into a more describing
'FakeVm', trivially fixing this issue.

Change-Id: I572a4f06a823e45088776c28f638ce5b249cbf7f
Test-Only: yes
Topic: master
Signed-off-by: Francesco Romani <fromani@redhat.com>