virt: Enable monitor response only if the VM is monitorable
libvirt operations in virdomain.Notifying set VM _monitorResponse flag
depending on whether operation timeout has or has not been experienced.
However the flag shouldn't be set to 0 when the VM is not monitorable.

For instance, when Vm.destroy is called and libvirt `destroy' operation fails
due to timeout (which has been actually observed) then the VM is left with
_monitorable set to False and _monitorResponse set to -1.  But there may follow
another libvirt call that doesn't time out, causing Vm._timeoutExperienced to
set _monitorResponse back to 0.  If the timeout and working libvirt calls
interval, _monitorResponse flips between -1 and 0.

However since the VM has _monitorable set to False, it's not actually monitored
and _monitorResponse should left _monitorResponse set -1 all the time, until
_monitorable is set to True and monitoring is recovered.  This patch prevents
_monitorResponse from being set to 0 if the VM is not monitorable.

Change-Id: I9806cfb6e3508b3528fdddf989822c995afd6820
Bug-Url: https://bugzilla.redhat.com/1481022
Backport-To: 4.2
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>