net: Update the timeout on a blocked-dhcp change
For blocking actions on ifup with dhcp, the ifup internal timeout is
arround 45sec. This change updates the waitfor call on waiting for an IP
address to 60sec.

Until this change, if the dhcp server responded after 10sec, the waitfor
would have already alerted on a timeout and even if the ifup would have
succeeded, it would not matter and the operation would fail on timeout.

Change-Id: I648726b925ea1527787f16ccb8444a20c07a8ef1
Signed-off-by: Edward Haas <edwardh@redhat.com>