tests: Remove unneeded context manager
Since we added the xtnd_message context manager, test_send_receive fails
randomly with this error:

00:08:21.810 2018-05-06 17:46:58,187 ERROR (mailbox-spm)
[storage.MailBox.SpmMailMonitor] SPM_MailMonitor: unknown message type
encountered: xtnd (mailbox:691)

I suspect that the test fails because we unregister the 'xtnd' message
just before the mail monitor thread check storage, creating a bogus
failure.

I did not invest the time to prove this theory, but having a context
manager that we don't need is a bad idea anyway; removing it.

Change-Id: If0462b8c43d043f4ae267ce5d4b6518c26b03de5
Signed-off-by: Nir Soffer <nsoffer@redhat.com>