virt: devices: do not remove data from kwargs
To properly support the VMFullList API in 4.1 compatibility mode,
we added code to device objects to track and report their parameters.

We need to avoid to remove keys from kwargs in
vmdevices.core.Base.__init__, to preserve a faithful copy of the
parameters given.

Previously, 'custom' may have got lost as unwanted side effect:
what code tried to do is to make sure it is present and initialized
to a meaningful value.

Change-Id: Id049ad2d99c13aa4ce450d457041c298ee7d94d1
Backport-To: 4.2
Topic: virt-tests
Signed-off-by: Francesco Romani <fromani@redhat.com>