virt: vm: hotunplug by alias
Add support of alias-only XML parameter in hotunplug flows of disks
and network interfaces.

With this patch, Engine can send a minimal XML specifying only the
alias of the device to be detached, and Vdsm will fill the gaps.
We expect libvirt to catch up "soon" and make the code in this patch
redundant.

The old hotunplug parameters are still supported.

Change-Id: Ia0b8f6e63afb3bb1cd03e087739012c2897d2524
Topic: hotunplug-alias
Backport-To: 4.2
Signed-off-by: Francesco Romani <fromani@redhat.com>
Signed-off-by: Arik Hadas <ahadas@redhat.com>