install: Prevent upgrades when QEMU is running
Although we don't support major version upgrades with VMs running on the host,
due to user errors VMs may actually be running on such an upgrade.  They are
not likely to survive the whole upgrade (including QEMU) without problems.  But
if VMs from pre-XML/metadata-recovery Vdsm version survive upgrade to
XML/metadata-recovery version at least partially, they may cause gross
confusion on the Engine side due to missing information, hard to repair.

There is no good solution to prevent the problem on the first upgrade.  But we
can introduce a yum plugin that prevents the problem on future upgrades.  This
patch introduces a yum plugin that aborts any Vdsm upgrade, reinstallation or
downgrade when QEMU processes are found.

Change-Id: I8905655e359f88ad86f35ad9e4e5af9b74827b66
Bug-Url: https://bugzilla.redhat.com/1570349
Backport-To: 4.2
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>