vm: make hotunplugMemory use new-style API
The current standard and best practice about implementing Vdsm API is
to let the implementation raise should any error occurr, or return
a dictionry of data to be returned in the response.

The API layer is in charge to build the response objects from either
values. In other words, the implementation should no longer return
response objects.

The implementation of VM.hotunplugMemory was half-baked: raising
exceptions on error, return responses on success. This works by
side-effect, and better avoiding that.

The simplest and most forward looking fix is to avoid returning
responses on success, and let the API layer do it.

Change-Id: Ib73f91c3d8cd5ee4b0ef70c2c76bc8828c94f073
Backport-To: 4.2
Topic: hotunplug-alias-2
Signed-off-by: Francesco Romani <fromani@redhat.com>