virt: metadata: ignore iotune in drive specParams
In the change Ifad7ed0dbbca1fdb61d740573f479804bb5fedbc
we started storing the drive specParams in the drives'
metadata to let the payload devices work across migrations.

The preliminary analysis of the patch explicitely considered
the impact of the change considering the IO tune settings.
Unfortunately the design of the patch was incorrect, because
the drive IO tune settings can end up in the specParams (old
design constraint), and the specParams must be serialized
- exactly to make the payload device work across migrations.

The simplest and safe fix is ignore any iotune setting in the
specParams when serializing or deserializing metadata.

This is correct because the premises of
Ifad7ed0dbbca1fdb61d740573f479804bb5fedbc
were right, it was the implementation at fault.

A future patch will streamline and simplify the storage of iotune
settings in the Drive object, because there are no real reasons,
besides historical constraints, to store it inside the specParams.

* 4.2 backport notice:
tests are been altered to reflect the fact that patch
Iab8d669db59aa084d6e1e730c700182c289b40e0 has not been backported yet
(lower priority)

Change-Id: I4ede28876b443541cf272b92833b535304c8e19a
Bug-Url: https://bugzilla.redhat.com/1579909
Backport-To: 4.2
Topic: metadata-drive-iotune
Signed-off-by: Francesco Romani <fromani@redhat.com>