vm: update DomainDescriptor after executing before_vm_start
Currently, the start flow with regards to DomainDescriptor looks like
this:
1. engine supplies XML definition to VDSM,
2. VDSM then creates initial DomainDescriptor from that definition,
3. before_vm_start hooks are executed against updated XML,
4. the domain is defined using updated and hook-modified XML,
5. the domain is started,
6. and, at last, we update the DomainDescriptor.

Now there is a huge issue if there is a hook that uses two specific
hooking points: before_vm_start and after_vm_destroy; and VM creation
fails. The XML changes done in before_vm_start hooking point are not
propagated to after_vm_destroy as, since creation failed, the
DomainDescriptor is not updated.

We can solve this problem by updating DomainDescriptor early.

Change-Id: Id084c2fe437131378d077c48fbb5225a3a90798b
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1581709
Signed-off-by: Martin Polednik <mpolednik@redhat.com>