net: Optimize DHCP info gathering in netinfo
DHCP information has been collected in two steps, once for the devices
(vlan, bond, nics, bridges) and once for the networks.

In some rare cases, the report included inconsistency between the
network dhcp state (fetched from the northbound iface) and the same
iface in the devices section.

It is suspected that the inconsistency appears due to the async nature
of the setup when dhcp is enabled.

Therefore, this patch suggests to optimize the dhcp info gathering to
occur in a single step and to execute it after the info is collected
from all the networks and devices.

Change-Id: Ia7113f4c4add08a15aba789f5bc6a250336a10f0
Bug-Url: https://bugzilla.redhat.com/1517638
Signed-off-by: Edward Haas <edwardh@redhat.com>