blockVolume: Remove metadata only if the LV successfully removed
Currently, if removing the LV fails, the volume is still returned by
image.getChain(), because it is not marked as _remove_me. In this case,
removing the metadata leads to expected exception when accessing the
invalid metadata. For example, if during live merge, removing the LV
fails but removing the metadata succeeds, when trying to create a
snapshot we got the following error:

MetaDataKeyNotFoundError: Meta Data key not found error:
("Missing metadata key: 'DOMAIN': found: {'NONE': '##########...

In this patch, we fix that behavior by removing the metadata only if
removing the LV succeeded.

Change-Id: I2cd771f9b21d43ca4cdd5295acdfbf1bfaf2882b
Bug-Url: https://bugzilla.redhat.com/1574631
Signed-off-by: Ala Hino <ahino@redhat.com>