sos: introduce dump-volume-chains option
This patch makes the collection of dump-volume-chains
optional. It can be used by the LogCollector to specify
a single host to run it, preventing duplicate data and
unecessary load on storage.

If ovirt-log-collector does not send this option, no
volume chains will be dumped. This is intentional
as at this point dumping chains is expensive and
often times out.

Usage:
$ sosreport -o vdsm -k vdsm.dump-volume-chains

This is to be used with the following change:
https://gerrit.ovirt.org/#/c/91979/

Change-Id: I0e6b4845485ec43a717881cf0dbd2678cbbe5e3f
Bug-Url: https://bugzilla.redhat.com/1489145
Signed-off-by: Germano Veit Michel <germano@redhat.com>