localfssd_test: add tests for LocalFsSD.createVolume()
Add tests for LocalFsSD.createVolume() in order to make
sure that after changing its logic
to use an initialSize and deferring the volume
preallocation the behavior will remain the same.

This patch is part of a series of patches that add the option to create
a raw preallocated volumes on file-based storage domain with an initial size.
This functionality is needed in order to keep the file volume format
when cloning it from a template and migrating it while the VM is down.

Change-Id: I9e0eb65e9e33b6a684b61ce13701bfd267e922a3
Bug-Url: https://bugzilla.redhat.com/1571285
Bug-Url: https://bugzilla.redhat.com/1403183
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>