vm: lease: on hotplug, ensure the 'device' key
When we hotplug a lease device, we store its params as the device
configuration in Vm.conf['devices']. Per API contract, the 'device'
parameter in the lease device configuration is not required.

When we call Vm._getUnderlyingInfo, most notably when we complete
a migration on destination side, the code inspects the Vm.conf[] content
to possibly upgrade the data based on the fresh XML from libvirt.

Device-specific logic may look for the 'device' key -to distinguish
the device configuration; this may lead to a KeyError, which ultimately
will make the migration fail. We can't recover, so the VM will end up
down.

This patch fixes this issue by making Vdsm always add this key to the
lease parameters it get. This is a little ugly, but always safe and
trivial to do, and ensures full compatibility with old Engines.

Change-Id: I6e3aaaa3ff627d65e6323121d9d88e06c70fe5ac
Bug-Url: https://bugzilla.redhat.com/1590063
Backport-To: 4.2
Topic: bz1590063
Signed-off-by: Francesco Romani <fromani@redhat.com>