vfio-mdev: move hook to codebase
Instead of having the vfio-mdev functionality available as a hook, we
move it to VDSMs codebase. The benefit of such approach is that we may
avoid future issues caused by the hooking mechanism and better
integration into VDSM itself (the hook already leaked into supervdsm).
This approach doesn't come without pitfalls either - mainly not
fitting into the new VM XML flow.

That being said, if we decide to support vfio-mdev natively from
engine, it is now much easier to simply remove XML and UUID generating
code and only keep the device handling code.

Change-Id: I423b419cd7305407e1bc3e879f73df3f3694a84a
Bug-Url: https://bugzilla.redhat.com/1581709
Backport-To: 4.2
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Signed-off-by: Martin Polednik <mpolednik@redhat.com>