pylint: disable bogus context manager warning
We are stuck with pylint 1.8, which fails to recognize some context
managers: see https://github.com/PyCQA/pylint/issues/782

Until we update to pylint 2.0+, silence warnings.

Unfortunately, we can't use anymore the -E option, because the settings
implied by that will undo the blacklisting of the buggy check.
The only option is be specific and toggle individual features.

Change-Id: Ia6e4fe92cbde7206cbd38363422ff118de486ae7
Topic: pylint-false-positives
Signed-off-by: Francesco Romani <fromani@redhat.com>