virt: Don't fail on missing <source> in a CD-ROM device
storagexml.parse tolerates missing <source> subelement and passes None to
_update_source_params in such a case.  However _update_source_params tries to
access `source' attributes in all the supported paths.

At least in case of CD-ROMs <source> element may be missing, after CD is
ejected.  If Vdsm is fenced in such a situation then VM recovery fails in
device initialization and the VM remains in Disconnected state internally.
That results in serious failures when Engine tries to handle the VM and may
result in split brain in case the VM is highly available.

This patch fixes the problem by handling None `source' value in
_update_source_params.

Change-Id: I7a76a2b317598828758948a717f5c12588e455fe
Bug-Url: https://bugzilla.redhat.com/1594793
Backport-To: 4.2
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>