imagetickets: Simpler and more correct reading
When reading response content, we used to get the content length and
read the exact amount of bytes promised.

Simplify the code to use HTTPResponse.read() to read the entire request
content. This ensure we support both response with a Content-Length
header, or response using chunked encoding.

This remove special handling for "204 No Content" as it looks the same
as "200 OK" with "Content-Length: 0".

Change-Id: I394c4e07d66225b28571ee0e34a87f9968bb0dce
Signed-off-by: Nir Soffer <nsoffer@redhat.com>