service: vdsm is dead after upgrade
When vdsm is upgraded together with libvirt due to dependency issue
vdsm process is not running. Moving dependency on libevirtd
from vdsm to supervdsm solves the issue. Vdsm depends on supervdsm so
the dependency is fulfilled indirectly.

We assume that this issue should be fixed by systemd and it is tracked
by BZ #1597179.


Bug-Url: https://bugzilla.redhat.com/1570562
Change-Id: Icfe67c83547d0a3a882d4d620efa88889e22eef8
Signed-off-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>