vm: simplify hasSpice
Let's always use the domain XML as authoritative source.

Please note that with this patch we have a time window
on which Vm.hasSpice return wrong results.
This spans from the moment the Vm is created to the moment
the domain XML is built, only for clusterLevel <= 4.1.

We use Vm.hasSpice only in the migration flow, and it is
extremely unlikely we manage to migrate before the domain XML
is built, so I believe this risk is negligible.

Change-Id: If6d1b35654662728e74f35f51c895f38f6469027
Topic: drop-4.1-compat
Breaks-Compatibility: 4.1
Signed-off-by: Francesco Romani <fromani@redhat.com>