virt: Retrieve VM memory info from system again
We previously replaced VM memory info retrieval with a libvirt call.  However
it occurred that the change is not without problems, since the libvirt call can
get stuck.  Not obtaining the corresponding information wouldn't be much
problem, but the stuck call is.  Among other it can cause monitorResponse value
flipping and Engine confusion.

For that reason we return back to the former method of VM memory info
retrieval, from the system.  Engine >= 4.2 doesn't use the provided information
so we can simply remove the whole VM info retrieval, which will happen in a
followup patch.  But we need to keep older Engine compatibility at least in
backport of this fix, so we need this patch at least for the backport.

Change-Id: I819981a55889f4e9a06e715d674bc48911aeef9e
Bug-Url: https://bugzilla.redhat.com/1481022
Backport-To: 4.2
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>