virt: Remove VM memory info retrieval
Engine >= 4.2 doesn't use memCommited and memAvailable values anymore.
So we can remove computation of those values.  However while Engine doesn't use
the values it still consumes them and we can't stop reporting them.  We can
just remove the machinery computing them and provide much simpler values (to be
on the safe side in comparison to setting them to 0 or so).

Change-Id: I4a93c34e89672ddc7c2461b676d00b56cc382671
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>