virt: Ignore device initialization errors in recovery
When a VM is being recovered and initialization of some device fails, Vm._dom
remains set to Disconnected.  This is a problem since the VM may be running and
Engine is unable to handle it.  That may lead to serious problems, up to split
brain.

That happened in https://bugzilla.redhat.com/1593568.  The VM was paused and
couldn't be unpaused due to Disconnected wrapper.  Then it couldn't be even
destroyed due to Disconnected wrapper.  Then it couldn't be started because it
already existed and when trying to destroy the newly started VM Engine was
tricked into believing the VM was down.  And started the VM elsewhere,
resulting in split brain.

This patch tries to make VM initialization on recovery more robust by ignoring
errors in device initialization from XML.  That means not all the devices may
be correctly initialized, but that's a better result of unexpected bugs than
not being able to handle the VM instance at all.

Change-Id: I2b56b77bee6b1d0ccb5a95498a75d17a43624624
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>