vm: network: handle disappearing defaults from XML
Libvirt recently started to NOT set anymore in the returned domain XML
the values which match the default settings.
IOW, if we explicitely set a value corresponding to a default, we may or
may not get the value back in the domain XML returned by libvirt,
depending on the libvirt version.

This is a legal behaviour, because implicit attributes have the default
value, so libvirt is actually pruning the unnecessary content from the
domain XML.

Also, it is not yet clear when libvirt started to do this, most likely
around version 3.9.0.

Another instance of this behaviour is the "queues" attribute in the
"driver" subelement of the interface device.
If we explicitely set queues=1, we don't get the attribute back when
querying libvirt. We instead get it when using any other value.

For historical reasons, Vdsm changes the value of the "driver" data
for device *objects* - not configuration - once reads back the domain
XML from libvirt. We need to preserve this behaviour to be compatible
with clusterLevel <= 4.1

In this patch, we *update* the value of the driver settings with data
from libvirt instead of replacing it.
The rationale for this change is to keep the data that may disappear
because it enforces a default, like queues=1.

I intentionally avoided fixes specific to the "queues" attribute, like
forcing the value to one if absent, because this looks too specific
and fragile, and I aimed instead for a more generic approach.

Change-Id: I7069c4eb9d698a2837a5b8f9713c2136fb12961d
Backport-To: 4.2
Topic: drop-4.1-compat
Signed-off-by: Francesco Romani <fromani@redhat.com>