vm: api: allow XML parameter in VM.updateDevice()
Bring VM.updateDevice() up to speed with the engine XML effort,
and add a 'xml' parameter, allowing Engine to send XML snippet.

So far this is implemented only for network device, because
the support is straightforward. To support also graphics devices,
we need to sort out how we will handle the free-form 'params'
as per schema.

Additionally, this patch adds the 'mtu' optional parameter to the
Interface class, to allow updates, or also to set the value at
creation time. This is actually needed, together with the XML support,
to fix the referenced BZ.

Change-Id: If599c9f6235870d1b3c770b9433e0a0e2344eec2
Topic: update-device-fix
Bug-Url: https://bugzilla.redhat.com/1600140
Signed-off-by: Francesco Romani <fromani@redhat.com>