vm: guestagent: store API version in the metadata
The guest agent API version is negotiated at runtime, once the
communication with guest agent starts.
Vdsm stores this value in the VM configuration to save time and
avoid the handshaking on recovery or migration.
Vdsm used to store the API version in the vm.conf, but with this
patch we store it only in the VM metadata.

Additionally, on migrations, Vdsm used to pass this value
to save the handshake on migration destination. We keep doing this,
making sure that the metadata has an up to date value when preparing
the migration.

To preserve the compatibility in the 4.1 to 4.1 migration,
we also copy the value from metadata - the authoritative source -
to the migration parameters.

Change-Id: I8e58f7dafc23831d8b615245bc5f53813da75046
Topic: drop-4.1-compat
Signed-off-by: Francesco Romani <fromani@redhat.com>