hsm: Limit qemu-img resources
Turns out qemu image parser is not hardened against malicious input and
can be abused to allocate an arbitrary amount of memory and/or dump a
lot of information when used with "--output=json".

The solution is to limit the resources used by qemu-img when using it
with untrusted image.

qemuimg.info() has now a new option "trusted_image", set to True by
default. Calling with trusted_image=False will limit the CPU time and
address space size of the underlying qemu-img command. The new option is
used when verifying untrusted images uploaded by users.

If "qemu-img info" to run more than 30 seconds or allocate more than 1
GiB, the command will terminate abnormally, failing the verification.

Here is an example run with a malicious image from the OpenStack bug:

$ /usr/bin/time -f "%MK" qemu-img info --output json afl3.img >/dev/null
1250976K

qemu-img tried to allocated 1.25G of memory.

When running with prlimit, it will fail with this error:

$ prlimit --as=1073741824 --cpu=1 qemu-img info --output json afl3.img
qemu-img: Could not open 'afl3.img': Cannot allocate memory

See the discussion in qemu-block mailing list:
http://lists.nongnu.org/archive/html/qemu-block/2018-07/msg00488.html

Original OpenStack bug:
http://lists.nongnu.org/archive/html/qemu-block/2018-07/msg00488.html

Change-Id: Idc810d1c1eab044353097944d98c3c9f0ea3da6e
Signed-off-by: Nir Soffer <nsoffer@redhat.com>