storage: add step=32M to blkdiscard (discard flow)
When discarding a device using blkdiscard without the 'step' flag,
'step' gets its default value, which is the size of the device.
Usually it functions fine, but in busy systems with certain storage
arrays, it can cause delays in other parallel IOs to the storage.

From the tests [1] that we ran, we found out that the optimal step size
is 32M, in terms of total discard time and minimal starvation of other
processes that run in parallel.
Therefore, to minimize potential starvation and enhance the performance
of blkdiscard, this patch adds a step=32M to blkdiscard calls that aim
to discard devices.

[1] To find the optimal step size, we created 8 * 100G pre-allocated
disks on a block domain, filled them with data and discarded all of them
in parallel from the same host.
At the same time, we ran the following loop to simulate reading from the
domain's metadata by sanlock (the loop was manually stopped when all the
discard calls ended ~ together):

for n in $(seq 1000); do
    echo "$(date +%T.%N) - read $n"
    time dd if=/path/to/domain/metadata of=/dev/null bs=1M count=1 iflag=direct
    sleep 1
done

The idea was to monitor the delays of each iteration, which is caused by
the starving the blkdiscard parallel calls.

Notes:
- The time it took for each iteration to complete when running without
the discard calls on the background was ~ 0.011s-0.014s.
- We calculated the minimum, maximum and average iteration time, as well
as the total discard time, for each step size: 16M, 32M, 64M, 128M, 256M
and with no step at all (which is essentially step=100G).

Here's a table that sums up the results on netapp:

Step    Average[s]   Min[s]   Max[s]   Total[min]
-------------------------------------------------
16M     0.013        0.011    0.025    6:33
32M     0.016        0.011    0.26     5:49
64M     0.14         0.011    1.72     5:37
128M    0.18         0.011    1.84     5:43
256M    0.35         0.019    1.8      5:58
100G    0.33         0.016    1.85     5:38

Here's a table of similar results on xtremio:

Step    Average[s]   Min[s]   Max[s]   Total[min]
-------------------------------------------------
32M     0.013        0.01     0.021    2:02
100G    4.89         0.54     13.59    2:32

Here're a few graphs that show the results:
https://imgur.com/a/PYO45uH

Change-Id: I5198115d4c9e6204399a43d81dd99c3e4d2902e7
Bug-Url: https://bugzilla.redhat.com/1562369
Signed-off-by: Idan Shaby <ishaby@redhat.com>