vm: devices: hotplug: don't update vm.conf
After hot{,un}plug, the Vm code used to update vm.conf['devices'],
which in turn was used to build getVMFullList responses and for
recovery.
For the latter, we depend on XML and libvirt already; for the former,
we have translation code to restore old style vm.conf parameters from
the domain XML, so we can get rid of the updating code.

Change-Id: If3a7b9b59aa7a570889d1f16cb1a97a996e14711
Topic: drop-4.1-compat
Signed-off-by: Francesco Romani <fromani@redhat.com>