api: change ctor for API.ISCSIConnection
As a workaround of the issue we want to change unusual format of the
ISCSIConnection object. We accpeted all params in ctor only to pass to
discoverSendTargets verb. In this patch we are changing to to match
other api objects.

This patch will have follow-up patches which will fix ordering and
validation for ctor params.


Bug-Url: https://bugzilla.redhat.com/1612958
Change-Id: I13362c21c8f28cba258c13060bdb6190308ff4ca
Signed-off-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>