virt: vm: Update guest time after VM un-pausing
Added optional Guest Time Synchronization to the snapshot functionality
via the time_sync_snapshot_enable option and other un-pausing scenarios
via the time_sync_cont_enable option for synchronizing and correcting
the time on the VM after long pauses.

This becomes especially critical when there are heavy loads on the VM to
ensure time stamps for example are accurate.

Change-Id: I9bf52cdcc15d15c37d723d81f0720a94aa6c65b5
Bug-Url: https://bugzilla.redhat.com/1510856
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>