net: Report correct mtu on the faked vlan iface
When OVS is used, a vlan is not defined using an interface, it is
defined by marking the OVS switch port as a vlan.
As Engine is not aware of this detail, the vlan interface is faked in
the caps network report.

In this patch, we set the network mtu on the faked vlan.

Change-Id: Iae48fd4011eafff9f1e65927a081cd3224a3bf4d
Bug-Url: https://bugzilla.redhat.com/1598461
Signed-off-by: Edward Haas <edwardh@redhat.com>