net: Checking if ipv6 is disabled, should not assume sysctl path
IPv6 stack may not be available on a specific interface, specifically
seen when an interface mtu is set to a value lower than 1280 (which is
the minimum mtu for IPv6).
When asked if IPv6 is enabled, it should be reasonable to answer if the
path for it is not found. This is also relevant when the node has
been loaded with IPv6 stack disabled from the boot line.

Change-Id: I762106ccc19317afcef57d29f97e0b63107f7381
Bug-Url: https://bugzilla.redhat.com/1598461
Signed-off-by: Edward Haas <edwardh@redhat.com>