net: Update the southbound iface mtu to the max mtus of all nets
The sb mtu should be set to the maximum value of the networks above it.
Specifically, the sb mtu should never be bigger than any of the networks mtus
define on top of it.

With this patch, the actual sb mtu is not considered when calculating
what mtu value to set on the sb iface.

Change-Id: I85fb474424a2ae1b7f22fc823d7ee38a55ad7bd9
Bug-Url: https://bugzilla.redhat.com/1598461
Signed-off-by: Edward Haas <edwardh@redhat.com>