virt: vm: Improved Flag Description for un-pausing
Improved Configuration Flag Descriptions for Guest Time Synchronization
of both the time_sync_snapshot_enable option and other un-pausing
scenarios via the time_sync_cont_enable option used for optionally
synchronizing and correcting the time on the VM after long pauses.

As per the descriptions, This becomes especially critical when there are
heavy loads on the VM to ensure time stamps for example are accurate.

Change-Id: I130786aef6334927e9928468df5d05094a521914
Bug-Url: https://bugzilla.redhat.com/1510856
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>