tests net: Add ip rules pre cleanup
The post cleanup of stale ip rules has been taking care of faulty tests
or issues that left stale ip rules on the node.
New faults have been detected where the existing identification of
these rules is itself faulty (BZ#1623488), requiring a different
approach.

As the existing cleanup is helpful in identifying bad tests
or other problems hiding under the hood, it is left as is and used as
post-cleanup. The new cleanup is added as pre-cleanup, running before
the tests start, cleaning stale ip rules which are expected to collide
with the tests generated ip rules.

Change-Id: Ibf807eb19d6d7223eb409faea8ae74b010381d21
Signed-off-by: Edward Haas <edwardh@redhat.com>