virt: Wait until migration is fully prepared before reporting success
On host-to-host migrations, the source host asks the destination host to create
migration, i.e. preparing the environment for starting the VM on the
destination.  The call waits until the migration is reported to be prepared and
if that succeeds, it reports success to the source host.

However the current code reports the success prematurely.  A success may be
reported and then VM initialization may fail in device setup, e.g. due to a bug
or a libvirt problem.  In such a case the VM is set to Down status on the
destination while the source host continues with the migration process.  If the
VM is set Down on the destination before the source actually starts the
migration (matter of timing), the migration proceeds without destination Vdsm
supervision.  libvirt migrates the VM while the destination Vdsm thinks there
is no VM there.  The destination Vdsm also rejects any VM destruction requests.

Since the VM runs on the destination stealthily, a split brain is going to
occur.

The remedy, implemented by this patch, is to wait with success report of
migrationCreate call until devices are initialized.  We reuse
Vm._pathsPreparedEvent for that purpose, since it is not used for any other
purpose these days.

Change-Id: I27d87f5cf0e03e6381da52943385ef10b9777e07
Bug-Url: https://bugzilla.redhat.com/1617745
Backport-To: 4.2
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>