virt: Catch NotConnectedError's in migration monitor thread
The migration monitor thread runs until it decides to stop or it is signaled to
stop.  It checks for self._stop event in its loop condition.

The problem is that the migration is finished first and only then stop is
called on the monitor thread.  When the migration is stopped due to VM
shutdown, crash or migration cancellation then it can happen that the monitor
thread operates on a no longer connected domain and crashes.  That has been
demonstrated in https://bugzilla.redhat.com/1572426.

While rare, it's still an expected situation so it shouldn't produce tracebacks
in the logs and the monitor thread should exit gracefully.  The easiest way to
ensure that is to catch NotConnectedError for the whole monitor_migration call.

Change-Id: I688c57268875b0a4f9c6919c3db824949732fe20
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>