virt: vm: Added Sparse support for Import
Added Sparseness support for Importing Virtual Machines when
Thin Provisioning is enabled and Volume or Streaming is enabled.

This is to ensure that imported images of Virtual Machines shall
retain the proper Disk Size and not the preallocated size when
Thin Provisioning is specified.

Note that Block Information specified as "path" in the code does
not support Sparseness so the implementation is limited to Volume
or Streaming.

Change-Id: I0eaa31ff212b73fe5b128595bfaf690e86cf5c04
Bug-Url: https://bugzilla.redhat.com/1587892
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>