virt: Log migration semaphore acquiring
In case of “stalled” migrations, it is useful to know exactly whether
migrations hang on that semaphore or on something else.  We used to provide a
debug-level log message about acquiring the semaphore, let's now use info-level
log messages to inform us when we hang on the semaphore and when we obtain it.

Change-Id: I665c8da842d790292895577b7bae2adf7a98546a
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>