py3: Fix 'Patch' for all class' methods' types
Patching with 'monkeypatch' in Python 3 is broken - all regular methods
were converted to static methods. This patch adds proper static method
detection and assertions to ensure, that all method calling semantics
remain valid after applying and reverting the patch.

Unfortunately there is no nicer way to check if a method is static
in Python 3, than to check a method's type taken directly from classes'
'__dict__'. When retrieving them with 'getattr', Python converts them
to functions and the information is gone. On the other hand, checking
if a method is a regular one, works differently for Python 2 and 3.
The 'inspect.ismethod' call returns true only for *bound* methods
in Python 3, thus a proper check would require creating an instance
of a class.

Change-Id: Ie8a01bc9ca13a5545c5da83dcee122853fb82a08
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>