virt: Fix mdev UUID check
When an mdev device is updated from XML, it looks for its alias among the
provided devices.  But the provided devices may be also host devices other than
MdevDevice's, not having `mdev_uuid' attribute, resulting in crash on the
attribute access.

This patch adds check for the particular host device type to prevent the error.

Change-Id: I85e3974dd1c8f81333493d239cfc0ec800c1a97d
Bug-Url: https://bugzilla.redhat.com/1629065
Backport-To: 4.2
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>