virt: utils: track flow in guestCpuLock
In order to better debug issues related to Vm._guestCpuLock,
we should track which flow holds the lock.
In this patch, we introduce a new helper class, a thin wrapper
around a threading.Lock() and reusing the same logic we had in
Vm._acquireCpuLockWithTimeout().

The advantage is that we can know which flow held the lock when
another flow tried -and failed- to get the lock itself, like in
the related-to BZ.

Moving the code in a helper class also improves the testability

Change-Id: Ic47b77bc55a88e867a27e71d26122cb216b84fef
Related-To: https://bugzilla.redhat.com/1620044
Topic: cpu-lock-timeout
Backport-To: 4.2
Signed-off-by: Francesco Romani <fromani@redhat.com>