api: Schema inconsistencies logging formatter
Introducing a custom-tailored formatter for log entries related to
schema inconsistencies. When logging with 'DEBUG' level, it adds
information gathered from stack to pinpoint specific issue.

This code is ugly, but extremely useful. Preferably, I would like
to get rid of it when most/all of the schema issues are resolved.

Example log entry:

Host.setKsmTune
With message: Required property pages_to_scan is not provided when calling Host.setKsmTune
With backtrace: [
	[
		"_verify_object_type",
		{
			"call_arg_keys":[
				"merge_across_nodes",
				"run"
			],
			"schema_type_name":"KsmTuneParams"
		}
	],
	[
		"_verify_complex_type",
		{
			"schema_type_type":"object"
		}
	]
]

Change-Id: I6dd00e29f188a81f3f206ebc360b9b8b840d80b5
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>