api: Schema inconsistencies logged at DEBUG level
The warnings regarding schema inconsistencies were disabled for both
production code and testing. The reason they were disabled
in production, is because they were bloating the logs significantly.
Messages emitted by the warnings were helpful, but not helpful enough
to resolve the issue quickly. This patch removes the warnings
completely and changes them to debug messages created by
'SchemaInconsistencyFormatter'. The 'schema.inconsistencies' logger will
be switched to 'DEBUG' level by OST. When enough issues are resolved,
'api_strict_mode' should be turned on and each inconsistency will
be treated as an error.

Change-Id: I081dfa423a27045834ddd5af0106c2d042b77348
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>