virt: Generalize device removal callback
onDeviceRemoved assumes that the callback is invoked only for explicitly
asynchronous device removals.  However it is invoked for all device removals
and results in warnings in the log such as

  Removed device not found in devices: scsi0-0-0-3

for hot unplugs of devices other than memory.

It's actually good that the callback is invoked for all device removals and we
can utilize that to improve synchronous device hot unplugs in Vdsm.  Let's
generalize the method to react properly to all device hot unplugs.  We'll
change synchronous device hot unplugs in followup patches.

Change-Id: I90e97eefe06749c2c784fa3a9f74ff6dc455de67
Bug-Url: https://bugzilla.redhat.com/1548846
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>